Use the same state chart for the pie and bar plots
In an effort to simplify the app logic, this commit refactors the pie
and bar plots to use the same state chart. The table chart has a similar
state chart, but is different enough that refactoring it would give no
real benefits.

Closes: #200

Squashed commits:
* refactor the bar chart machine
* refactor the pie chart machine
