+++ src/components/QueryController/QueryController.jsx
@@ -3,7 +3,11 @@ import { IconNames } from '@blueprintjs/icons'
 import { useService } from '@xstate/react'
 import PropTypes from 'prop-types'
 import React from 'react'
-import { DELETE_OVERVIEW_CONSTRAINT_FROM_QUERY, FETCH_OVERVIEW_SUMMARY } from 'src/eventConstants'
+import {
+	DELETE_OVERVIEW_CONSTRAINT_FROM_QUERY,
+	RESET_PLOTS,
+	UPDATE_OVERVIEW_PLOTS,
+} from 'src/eventConstants'
 import { QueryServiceContext, useEventBus } from 'src/useEventBus'
 
 import { CODES } from '../common'
@@ -111,7 +115,8 @@ export const QueryController = ({ queryControllerActor }) => {
 			where: codedConstraints,
 		}
 
-		sendToBus({ type: FETCH_OVERVIEW_SUMMARY, query, classView, rootUrl })
+		sendToBus({ type: RESET_PLOTS })
+		sendToBus({ type: UPDATE_OVERVIEW_PLOTS, query, classView, rootUrl })
 	}
 
 	const handleDeleteConstraint = (path) => {
